pyproject.toml
[tool.poetry]
name = "Sidewinder"
version = "0.1.0"
description = "Django starter kit"
authors = ["Petr Stribny <petr@stribny.name>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "4.0"
djangorestframework = "3.13.1"
Markdown = "3.3.6"
django-filter = "21.1"
django-extensions = "3.1.5"
rich = "11.0.0"
Faker = "11.3.0"
django-allauth = "0.48.0"
django-bulma = "0.8.3"
argon2-cffi = "21.3.0"
django-htmx = "^1.9.0"
drf-spectacular = "^0.21.2"
drf-spectacular-sidecar = "^2022.3.14"
pendulum = "^2.1.2"
django-environ = "^0.9.0"
dj-database-url = "^1.0.0"
Pillow = "^9.2.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.3"
django-structlog = "^3.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
django-browser-reload = "1.3.0"
django-debug-toolbar = "3.2.4"
pytest-django = "4.5.2"
pytest-icdiff = "0.5"
pytest-randomly = "3.11.0"
pytest-cov = "3.0.0"
pywatchman = "1.4.1"
pytest-sugar = "0.9.4"
bandit = "1.7.3"
black = "22.3.0"
flake8 = "4.0.1"
snoop = "0.4.1"
pre-commit = "2.17.0"
playwright = "1.19.1"
openapi-spec-validator = "^0.4.0"
poethepoet = "^0.16.0"
django-stubs = {version = "^1.12.0", extras = ["compatible-mypy"]}
django-silk = "^5.0.1"
pytest-factoryboy = "^2.5.0"
isort = "^5.10.1"
djangorestframework-stubs = {extras = ["compatible-mypy"], version = "^1.7.0"}
mkdocs = "^1.4.2"
mkdocs-material = "^8.5.11"
pydotplus = {version = "^2.0.2", optional = true}
pygraphviz = {version = "^1.10", optional = true}

[tool.poe.tasks]
ma = "./manage.py"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
